Add Cinch::User#blolol
Cinch::User#blolol returns an instance of Wheaties::BlololUser, which
talks to Blolol's API and makes Blolol user details available in a more
convenient way than the `blolol_user_info` Wheaties command. It also
makes an attempt to work with RELAYMSG users, such as those chatting in
Discord via Matterbridge. (In those cases, Wheaties attempts to look up
the user based on the Discord nick, which might not always match.)