Return string instead of single-element array
The change in 302f30f to make plain text and random line commands always
return an array of strings broke an assumption made in many existing
Ruby commands that both types of text commands would always return
a string. This commit changes their behavior so that, if the sorting and
filtering options applied result in a single line, then just that one
string is returned. Otherwise, an array of strings is returned as before.